+++ functions/backend/update_system.php
@@ -12,7 +12,7 @@
         if($isNewer && $_SESSION['admin'])
         {
             $updateBtn = '
-            <div class="shrink-0">
+            
                 <button type="button" id="update-btn" class="relative inline-flex items-center gap-x-1.5 bg-sky-600 px-3 py-2 text-sm font-semibold text-white shadow-xs hover:bg-sky-500 focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-sky-600">
                     <svg class="-ml-0.5 w-5 h-5" viewBox="0 0 25 25" fill="currentColor" aria-hidden="true" data-slot="icon">
                     <path d="M12.5535 16.5061C12.4114 16.6615 12.2106 16.75 12 16.75C11.7894 16.75 11.5886 16.6615 11.4465 16.5061L7.44648 12.1311C7.16698 11.8254 7.18822 11.351 7.49392 11.0715C7.79963 10.792 8.27402 10.8132 8.55352 11.1189L11.25 14.0682V3C11.25 2.58579 11.5858 2.25 12 2.25C12.4142 2.25 12.75 2.58579 12.75 3V14.0682L15.4465 11.1189C15.726 10.8132 16.2004 10.792 16.5061 11.0715C16.8118 11.351 16.833 11.8254 16.5535 12.1311L12.5535 16.5061Z"/> 
@@ -20,9 +20,9 @@
                     </svg>
                     Update available
                 </button>
-            </div>';
+            ';
             $updateDockerBtn = '
-            <div class="shrink-0">
+            
                 <button type="button" id="update-btn-docker" class="relative inline-flex items-center gap-x-1.5 bg-sky-500 px-3 py-2 text-sm font-semibold text-white shadow-xs hover:bg-sky-600 focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-sky-600">
                     <svg class="-ml-0.5 w-5 h-5" viewBox="0 0 25 25" fill="currentColor" aria-hidden="true" data-slot="icon">
                     <path d="M12.5535 16.5061C12.4114 16.6615 12.2106 16.75 12 16.75C11.7894 16.75 11.5886 16.6615 11.4465 16.5061L7.44648 12.1311C7.16698 11.8254 7.18822 11.351 7.49392 11.0715C7.79963 10.792 8.27402 10.8132 8.55352 11.1189L11.25 14.0682V3C11.25 2.58579 11.5858 2.25 12 2.25C12.4142 2.25 12.75 2.58579 12.75 3V14.0682L15.4465 11.1189C15.726 10.8132 16.2004 10.792 16.5061 11.0715C16.8118 11.351 16.833 11.8254 16.5535 12.1311L12.5535 16.5061Z"/> 
@@ -30,7 +30,7 @@
                     </svg>
                     New Docker Image available
                 </button>
-            </div>';
+            ';
         }
 
         if(!$docker)